Move SharedAsset[Depot] to CesiumAsync.
