+++ CesiumGltf/include/CesiumGltf/ImageCesium.h
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "CesiumAsync/SharedAsset.h"
 #include "CesiumGltf/Ktx2TranscodeTargets.h"
 #include "CesiumGltf/Library.h"
-#include "CesiumGltf/SharedAsset.h"
 
 #include <cstddef>
 #include <cstdint>
@@ -30,7 +30,8 @@ struct CESIUMGLTF_API ImageCesiumMipPosition {
  * @brief Holds {@link Image} properties that are specific to the glTF loader
  * rather than part of the glTF spec.
  */
-struct CESIUMGLTF_API ImageCesium final : public SharedAsset<ImageCesium> {
+struct CESIUMGLTF_API ImageCesium final
+    : public CesiumAsync::SharedAsset<ImageCesium> {
   /**
    * @brief The width of the image in pixels.
    */
